+++ fabric/src/main/java/io/github/sakurawald/fuji/module/initializer/launcher/LauncherInitializer.java
@@ -1,58 +1,113 @@
 package io.github.sakurawald.fuji.module.initializer.launcher;
 
-import io.github.sakurawald.fuji.core.auxiliary.LogUtil;
 import io.github.sakurawald.fuji.core.auxiliary.minecraft.CommandHelper;
-import io.github.sakurawald.fuji.core.auxiliary.minecraft.EntityHelper;
 import io.github.sakurawald.fuji.core.command.annotation.CommandNode;
 import io.github.sakurawald.fuji.core.command.annotation.CommandRequirement;
 import io.github.sakurawald.fuji.core.command.annotation.CommandSource;
 import io.github.sakurawald.fuji.core.command.argument.wrapper.impl.EntityCollection;
+import io.github.sakurawald.fuji.core.document.annotation.ColorBox;
 import io.github.sakurawald.fuji.core.document.annotation.Document;
-import io.github.sakurawald.fuji.core.event.impl.ServerTickEvents;
-import io.github.sakurawald.fuji.core.manager.Managers;
-import io.github.sakurawald.fuji.core.manager.impl.task.structure.GameTask;
 import io.github.sakurawald.fuji.module.initializer.ModuleInitializer;
-import java.util.ArrayList;
-import java.util.List;
+import java.util.HashSet;
+import java.util.Set;
 import net.minecraft.entity.Entity;
 import net.minecraft.server.command.ServerCommandSource;
 import net.minecraft.util.math.MathHelper;
-import net.minecraft.util.math.Vec3d;
+import org.jetbrains.annotations.NotNull;
 
 
 @Document(id = 1756347408511L, value = """
-    Launch an `entity` with a `velocity`.
+    Launch a target `entity` in specified `direction` and `power`.
+    """)
+@ColorBox(id = 1756440410476L, color = ColorBox.ColorBoxTypes.EXAMPLE, value = """
+    ◉ Use a lower `angle` for `fast horizontal movement`.
+    1. `/launch facing @s 15 1`
+    2. `/launch facing @s 15 3.9`
+    <green>The `power` is clamped in `[-3.9, +3.9]`
+
+    ◉ Use a median `angle` for `balanced horizontal and vertical movement`.
+    1. `/launch facing @s 30 3.9`
+    2. `/launch facing @s 45 3.9`
+
+    ◉ Use a higher `angle` for a `rocket launcher` effect.
+    Issue: `/launch facing @s 60 3.9`
+    <green>TIP: Remember to bring your `elytra`.
+
+    ◉ Use a `vertical angle` for a `trampoline` effect.
+    Issue: `/launch facing @s 90 1`
+
+    ◉ Use a `positive power` for a `push` effect.
+    Issue: `/launch facing @s 0 1`
+
+    ◉ Use a `negative power` for a `pull` effect.
+    Issue: `/launch facing @s 0 -1`
+
+    ◉ Use another entity's perspective as the direction, to `kick` the target entity.
+    Issue: `/launch at @s @e[type=!minecraft:player,distance=..8] 30 1`
+
+    ◉ Create a `jump pad` that launches players when stepped on.
+    You can integrate with `command_attachment` module.
+    Issue: `/command-attachment attach-block-one ~ ~ ~ --interactType STEP_ON \\<command\\>`
     """)
 public class LauncherInitializer extends ModuleInitializer {
 
-//    // NOTE: The base velocity for walking and flying is different.
-//    private static final double BLOCKS_PER_TICK = 2.2010;
-//    // In vanilla Minecraft, the max distance an entity can go through in 20 ticks is 82.184 blocks.
-//    // For a still player, the gravity acceleration is (0.0, -0.0784000015258789, 0.0). If the player is flying, then it's (0, 0, 0)
-//    // The max distance an entity can go through in 1 tick is 9.792
-//    // 8.084
-//
-//    @CommandNode("launch")
-//    @CommandRequirement(level = 4)
-//    private static int $launch(@CommandSource ServerCommandSource source, EntityCollection target, double x, double y, double z, int ticks) {
-//        List<GameTask> createdTasks = new ArrayList<>();
-//        target
-//            .getValue()
-//            .forEach(entity -> {
-//                GameTask gameTask = new GameTask(ticks, () -> {
-//                    LogUtil.warn("onTick() -> velocity = {}", entity.getVelocity());
-//                    EntityHelper.setVelocity(entity, x, y, z);
-//                }, () -> {
-//                    LogUtil.warn("onStart() -> velocity = {}", entity.getVelocity());
-//                }, () -> {
-//                    LogUtil.warn("onEnd() -> velocity = {}", entity.getVelocity());
-//                });
-//                createdTasks.add(gameTask);
-//            });
-//
-//        Managers.getGameTaskManager().submitTasks(createdTasks);
-//        return CommandHelper.Return.SUCCESS;
-//    }
+    public static Set<Entity> LAUNCHED_ENTITIES = new HashSet<>();
+
+    private static void launchEntityFacing(@NotNull Entity player, float angle, double power) {
+        launchEntity(player, player.getYaw(), angle, power);
+    }
+
+    private static void launchEntityAt(@NotNull Entity player, @NotNull Entity at, float angle, double power) {
+        launchEntity(player, at.getYaw(), angle, power);
+    }
+
+    public static void launchEntity(@NotNull Entity entity, float perspectiveYaw, float perspectivePitch, double power) {
+        /* Compute yaw and pitch. */
+        float yaw = perspectiveYaw * ((float) Math.PI / 180F);
+        float pitch = perspectivePitch * ((float) Math.PI / 180F);
+
+        /* Compute directional vector from yaw + pitch. */
+        double x = -Math.sin(yaw) * Math.cos(pitch);
+        double y = Math.sin(pitch);
+        double z = Math.cos(yaw) * Math.cos(pitch);
+
+        /* Normalize. */
+        double length = Math.sqrt(x * x + y * y + z * z);
+        x /= length;
+        y /= length;
+        z /= length;
+
+        /* Apply velocity scaled by power. */
+        power = MathHelper.clamp(power, -3.9F, 3.9F);
+        entity.setVelocity(x * power, y * power, z * power);
+
+        /* Mark velocity as modified. */
+        entity.velocityModified = true;
+    }
+
+    @CommandNode("launch facing")
+    @CommandRequirement(level = 4)
+    private static int $launch(@CommandSource ServerCommandSource source, EntityCollection target, float angle, double power) {
+        target
+            .getValue()
+            .forEach(entity -> {
+                LAUNCHED_ENTITIES.add(entity);
+                launchEntityFacing(entity, angle, power);
+            });
+        return CommandHelper.Return.SUCCESS;
+    }
+
+    @CommandNode("launch at")
+    @CommandRequirement(level = 4)
+    private static int $launch(@CommandSource ServerCommandSource source, Entity at, EntityCollection target, float angle, double power) {
+        target
+            .getValue()
+            .forEach(entity -> {
+                LAUNCHED_ENTITIES.add(entity);
+                launchEntityAt(entity, at, angle, power);
+            });
+        return CommandHelper.Return.SUCCESS;
+    }
 
 
 }